moved CorrectionPopinProps
