+++ packages/@coorpacademy-components/src/organism/review-stacked-slides/test/fixtures/all-ok.ts
@@ -1,5 +1,5 @@
 import AnswerQCMDrag from '../../../../molecule/answer/test/fixtures/qcm-drag';
-import {CorrectionPopinProps} from '../../../review-slide/prop-types';
+import {Props as CorrectionPopinProps} from '../../../../molecule/review-correction-popin/prop-types';
 import {Props} from '../../prop-types';
 
 type Fixture = {
