+++ packages/@coorpacademy-components/src/organism/review-stacked-slides/test/fixtures/unstack.ts
@@ -1,6 +1,6 @@
 import AnswerQCMDrag from '../../../../molecule/answer/test/fixtures/qcm-drag';
 import AnswerQCMGraphic from '../../../../molecule/answer/test/fixtures/qcm-graphic';
-import {CorrectionPopinProps} from '../../../review-slide/prop-types';
+import {Props as CorrectionPopinProps} from '../../../../molecule/review-correction-popin/prop-types';
 import {Props} from '../../prop-types';
 
 type Fixture = {
